[discover][table] Move update's discover logic to separate class
